backend/auth-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic[email]==2.5.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
